Assume only C90, so cast to unsigned long for size_t formatting

+++ utils/utils.h
@@ -28,21 +28,21 @@
 #define MALCHOKE(pChunk,chunkSize,chunkCast)				\
   if ((pChunk = (chunkCast) malloc(chunkSize)) == 0) {			\
     fprintf (stderr,"malloc of %lu bytes for variable %s cast as (%s) failed at %s:%d!\n", \
-	     chunkSize,#pChunk,#chunkCast,(__FILE__),(__LINE__));	\
+	     (unsigned long)chunkSize,#pChunk,#chunkCast,(__FILE__),(__LINE__));	\
     exit (EXIT_FAILURE);						\
   }
 
 #define REALCHOKE(pChunk,chunkSize,chunkCast)				\
   if ((pChunk = (chunkCast) realloc(pChunk, chunkSize)) == 0) {		\
     fprintf (stderr,"realloc to %lu bytes of variable %s cast as (%s) failed at %s:%d!\n", \
-	     chunkSize,#pChunk,#chunkCast,(__FILE__),(__LINE__));	\
+	     (unsigned long)chunkSize,#pChunk,#chunkCast,(__FILE__),(__LINE__));	\
     exit (EXIT_FAILURE);						\
   }
 
 #define CALCHOKE(pChunk,chunkSize,chunkCount,chunkCast)			\
   if ((pChunk = (chunkCast) calloc(chunkSize, chunkCount)) == 0) {	\
     fprintf (stderr,"calloc of %lu * %lu zeroed bytes of variable %s cast as (%s) failed at %s:%d!\n", \
-	     chunkSize,chunkCount,#pChunk,#chunkCast,(__FILE__),(__LINE__)); \
+	     (unsigned long)chunkSize,(unsigned long)chunkCount,#pChunk,#chunkCast,(__FILE__),(__LINE__)); \
     exit (EXIT_FAILURE);						\
   }
